+++ core/base/src/commonDomMain/kotlin/nl/adaptivity/xmlutil/util/CompactFragment.kt
@@ -59,7 +59,7 @@ public actual class CompactFragment : ICompactFragment {
 
     actual override fun serialize(out: XmlWriter) {
         XMLFragmentStreamReader.from(this).use { reader ->
-            out.serialize(reader)
+            out.serialize(reader, keepDuplicatedNsDecls = false)
         }
     }