+++ core/base/src/commonMain/kotlin/nl/adaptivity/xmlutil/util/CompactFragmentSerializer.kt
@@ -1,21 +1,21 @@
 /*
- * Copyright (c) 2024.
+ * Copyright (c) 2024-2025.
  *
  * This file is part of xmlutil.
  *
- * This file is licenced to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You should have received a copy of the license with the source distribution.
- * Alternatively, you may obtain a copy of the License at
+ * This file is licenced to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance
+ * with the License.  You should have  received a copy of the license
+ * with the source distribution. Alternatively, you may obtain a copy
+ * of the License at
  *
  *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
  */
 
 package nl.adaptivity.xmlutil.util
@@ -32,6 +32,7 @@ public object CompactFragmentSerializer : XmlSerializer<CompactFragment> {
     private val namespacesSerializer = ListSerializer(Namespace)
 
     override val descriptor: SerialDescriptor = buildClassSerialDescriptor("nl.adaptivity.xmlutil.util.compactFragment") {
+        annotations = listOf(XmlDynamicNameMarker())
         element("namespaces", namespacesSerializer.descriptor)
         element("content", serialDescriptor<String>())
     }